refactor: move functions out of AFEP_parse and into fepruns.py
